+++ src/ResourceManager/Compute/Commands.Compute/Properties/Resources.Designer.cs
@@ -297,6 +297,24 @@ namespace Microsoft.Azure.Commands.Compute.Properties {
         }
         
         /// <summary>
+        ///   Looks up a localized string similar to Enable AzureDiskEncryption on the VM.
+        /// </summary>
+        public static string EnableAzureDiskEncryptionCaption {
+            get {
+                return ResourceManager.GetString("EnableAzureDiskEncryptionCaption", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Looks up a localized string similar to This cmdlet prepares the VM and enables encryption which may reboot the machine and takes 10-15 minutes to finish. Please save your work on the VM before confirming. Do you want to continue?.
+        /// </summary>
+        public static string EnableAzureDiskEncryptionConfirmation {
+            get {
+                return ResourceManager.GetString("EnableAzureDiskEncryptionConfirmation", resourceCulture);
+            }
+        }
+        
+        /// <summary>
         ///   Looks up a localized string similar to Error occurred when creating storage account for boot diagnostics.  Keep creating a VM with disabling boot diagnostics.  : {0}.
         /// </summary>
         public static string ErrorDuringCreatingStorageAccountForBootDiagnostics {
@@ -928,6 +946,15 @@ namespace Microsoft.Azure.Commands.Compute.Properties {
         public static string VirtualMachineStoppingConfirmation {
             get {
                 return ResourceManager.GetString("VirtualMachineStoppingConfirmation", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Looks up a localized string similar to You have to specify either both of KeyEncryptionKeyVaultId and KeyEncryptionKeyUrl or none of them..
+        /// </summary>
+        public static string VMOSDiskDiskEncryptionBothKekVaultIdAndKekUrlRequired {
+            get {
+                return ResourceManager.GetString("VMOSDiskDiskEncryptionBothKekVaultIdAndKekUrlRequired", resourceCulture);
             }
         }
     }